+++ packages/peregrine/lib/util/__tests__/getCurrencyCode.spec.js
