+++ packages/peregrine/lib/util/getCurrencyCode.js
