+++ packages/venia-ui/lib/components/MiniCart/miniCart.js
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from 'react';
+import React from 'react';
 import { bool, shape, string } from 'prop-types';
 
 import Body from './body';
@@ -8,75 +8,31 @@ import Mask from './mask';
 import defaultClasses from './miniCart.css';
 
 import { mergeClasses } from '../../classify';
-import getCurrencyCode from '../../util/getCurrencyCode';
-import { useAppContext } from '@magento/peregrine/lib/context/app';
-import { useCartContext } from '@magento/peregrine/lib/context/cart';
-import { useCheckoutContext } from '@magento/peregrine/lib/context/checkout';
+import { useMiniCart } from '@magento/peregrine/lib/talons/MiniCart/useMiniCart';
 
 const MiniCart = props => {
-    const [, { closeDrawer }] = useAppContext();
-    const [
+    const {
+        cartItems,
         cartState,
-        { updateItemInCart, removeItemFromCart }
-    ] = useCartContext();
-    const [, { cancelCheckout }] = useCheckoutContext();
-    const [step, setStep] = useState('cart');
+        currencyCode,
+        handleBeginEditItem,
+        handleDismiss,
+        handleEndEditItem,
+        handleClose,
+        handleUpdateItemInCart,
+        isEditingItem,
+        isLoading,
+        isMiniCartMaskOpen,
+        isUpdatingItem,
+        numItems,
+        removeItemFromCart,
+        setStep,
+        shouldShowFooter,
+        step,
+        subtotal
+    } = useMiniCart();
 
-    // Props.
-    const { isOpen } = props;
-
-    const { isLoading, isUpdatingItem } = cartState;
-
-    const [isEditingItem, setIsEditingItem] = useState(false);
-
-    // Members.
-    const classes = mergeClasses(defaultClasses, props.classes);
-    const currencyCode = getCurrencyCode(cartState);
-    const cartItems = cartState.details.items;
-    const numItems = cartState.details.items_qty;
-    const rootClass = isOpen ? classes.root_open : classes.root;
-    const subtotal = cartState.totals.subtotal;
-
-    const showFooter =
-        step === 'receipt' ||
-        step === 'form' ||
-        !((cartState.isEmpty && step === 'cart') || isLoading || isEditingItem);
-
-    const isMiniCartMaskOpen = step === 'form';
-
-    const handleClose = useCallback(() => {
-        setStep('cart');
-        setIsEditingItem(false);
-        closeDrawer();
-    }, [closeDrawer, setStep]);
-
-    const handleBeginEditItem = useCallback(() => {
-        setIsEditingItem(true);
-    }, []);
-
-    const handleEndEditItem = useCallback(() => {
-        setIsEditingItem(false);
-    }, []);
-
-    const handleUpdateItemInCart = useCallback(
-        async (...args) => {
-            try {
-                await updateItemInCart(...args);
-            } catch (error) {
-                console.log('Unable to update item:', error.message);
-            } finally {
-                setIsEditingItem(false);
-            }
-        },
-        [updateItemInCart]
-    );
-
-    const handleDismiss = useCallback(() => {
-        setStep('cart');
-        cancelCheckout();
-    }, [cancelCheckout]);
-
-    const footer = showFooter ? (
+    const footer = shouldShowFooter ? (
         <Footer
             currencyCode={currencyCode}
             isMiniCartMaskOpen={isMiniCartMaskOpen}
@@ -86,6 +42,9 @@ const MiniCart = props => {
             subtotal={subtotal}
         />
     ) : null;
+
+    const classes = mergeClasses(defaultClasses, props.classes);
+    const rootClass = props.isOpen ? classes.root_open : classes.root;
 
     return (
         <aside className={rootClass}>
